% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel}
\alias{excel}
\title{excel}
\usage{
excel(
  dfs = null,
  file_name = null,
  folder = null,
  .debug = T,
  env = rlang::caller_env(),
  ...
)
}
\arguments{
\item{dfs}{object or list of data frame to write}

\item{file_name}{file name to save}

\item{folder}{folder to save file}

\item{.debug}{for internal use}

\item{env}{environment}

\item{...}{for future versions}
}
\value{
object or list of data frame back
}
\description{
excel
}
\examples{
\dontrun{
excel( data.frame(a = 1 : 3 ) , file_name = "test1.xlsx" , folder = "." )
}
}
