% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_df_utils.R
\name{remove_na_safe}
\alias{remove_na_safe}
\title{remove_na_safe}
\usage{
remove_na_safe(df, verbose = F)
}
\arguments{
\item{df}{data.frame to remove na rows from the beginning and from the end}

\item{verbose}{give detailed info while removing NA values}
}
\description{
remove_na_safe
}
